Move `test_download`
